Rewrite handling of the fixed rates
Make srate signed, or checks for it being less than 0 don't work.
Consistently use IEEE80211_FIXED_RATE_NONE.  Fix misleading comments.

Survive the case when the fixed rate is not in the rate table.  It can
happen e.g. a CCK rate is used with an 802.11a channel, or in the STA
mode when an OFDM rate is used with 802.11b.  The driver should catch
it, but it doesn't.  In any case, it's not a reason for an oops.

Remove the debugging message if everything is fine.  In case of an
error, specify the VAP name.